README: Mention AnsiToImg as an alternative tool
I found another tool that is similar in some ways. It is based on the Python `rich` library and is less flexible, but it impressed me in that the SVGs worked with Inkscape and browsers out of the box. I'm not sure if it's still the case after https://github.com/slowli/term-transcript/pull/53, but perhaps there are some tricks that you could borrow from it.